requirements.txt
matplotlib
numpy
pydantic
pyshtools
scipy
watchdog
pylint
geopandas
shapely